changed data types of Prices

+++ database/migrations/2024_03_31_215709_create_services_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('services', function (Blueprint $table) {
             $table->id()->primary();
             $table->string("name");
-            $table->float("price");
+            $table->decimal('price',8,2);
             $table->timestamps();
         });
     }
